ms-asr-model/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ms-asr-model"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.62.0",
    "grpcio-tools>=1.62.0",
    "faster-whisper>=1.0.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.26.0",
]

[tool.setuptools.packages.find]
where = ["src"]
